Define the SpatialCollection ADT
What: Create a new ADT, Spatial
Collection to describe the behavior of a
data structure containing objects that are
spatially distributed.
Why: The spatial distribution of vaccines'
warehouses is a core piece of the
problem that will be solved. Also, this
abstraction allows to use the Strategy
Pattern to change the actual data
structure more easily